pyproject.toml
[project]
name = "digital-card"
version = "0.1.0"
description = "Tap-to-Wallet FastAPI relay (WalletWallet)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2024.0.0",
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.46.0",
    "requests>=2.33.1",
    "python-dotenv>=1.2.2",
    "pymongo>=4.6",
]

[project.optional-dependencies]
dev = ["httpx>=0.28.1"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[build-system]
requires = ["setuptools>=82.0.1"]
build-backend = "setuptools.build_meta"
